Fix 952 Feature
