+++ public/locales/ar.json
@@ -171,7 +171,8 @@
       "confirmDelete": "هل أنت متأكد من رغبتك في حذف هذا الحدث؟",
       "delete": "حذف",
       "manageRegistrations": "إدارة التسجيلات",
-      "noGroups": "لم يتم العثور على مجموعات."
+      "noGroups": "لم يتم العثور على مجموعات.",
+      "editEvent": "تحرير الحدث"
     },
     "calendarList": {
       "aboutParagraph1": "تتيح لك التقاويس تنظيم وعرض الأحداث في كنيستك.",
