+++ public/locales/hi.json
@@ -171,7 +171,8 @@
       "confirmDelete": "क्या आप वाकई इस इवेंट को हटाना चाहते हैं?",
       "delete": "हटाएं",
       "manageRegistrations": "पंजीकरण प्रबंधित करें",
-      "noGroups": "कोई समूह नहीं मिला।"
+      "noGroups": "कोई समूह नहीं मिला।",
+      "editEvent": "ईवेंट संपादित करें"
     },
     "calendarList": {
       "aboutParagraph1": "कैलेंडर आपको अपने चर्च के लिए कार्यक्रम आयोजित करने और प्रदर्शित करने की अनुमति देते हैं।",
