+++ public/locales/ko.json
@@ -171,7 +171,8 @@
       "confirmDelete": "이 이벤트를 삭제하시겠습니까?",
       "delete": "삭제",
       "manageRegistrations": "등록 관리",
-      "noGroups": "그룹을 찾을 수 없습니다."
+      "noGroups": "그룹을 찾을 수 없습니다.",
+      "editEvent": "이벤트 편집"
     },
     "calendarList": {
       "aboutParagraph1": "달력을 사용하면 교회의 행사를 정리하고 표시할 수 있습니다.",
