+++ public/locales/ru.json
@@ -171,7 +171,8 @@
       "confirmDelete": "Вы уверены, что хотите удалить это событие?",
       "delete": "Удалить",
       "manageRegistrations": "Управление регистрациями",
-      "noGroups": "Группы не найдены."
+      "noGroups": "Группы не найдены.",
+      "editEvent": "Редактировать событие"
     },
     "calendarList": {
       "aboutParagraph1": "Календари позволяют вам организовывать и отображать события для вашей церкви.",
