+++ public/locales/sk.json
@@ -171,7 +171,8 @@
       "confirmDelete": "Are you sure you wish to delete this event?",
       "delete": "Vymazať",
       "manageRegistrations": "Spravovať registrácie",
-      "noGroups": "No groups found."
+      "noGroups": "No groups found.",
+      "editEvent": "Upraviť udalosť"
     },
     "calendarList": {
       "aboutParagraph1": "Calendars allow you to organize and display events for your church.",
