+++ public/locales/sr.json
@@ -171,7 +171,8 @@
       "confirmDelete": "Да ли сте сигурни да желите да обришете овај догађај?",
       "delete": "Обриши",
       "manageRegistrations": "Управљај регистрацијама",
-      "noGroups": "Нема пронађених група."
+      "noGroups": "Нема пронађених група.",
+      "editEvent": "Uredi događaj"
     },
     "calendarList": {
       "aboutParagraph1": "Календари вам омогућавају да организујете и приказујете догађаје ваше цркве.",
