+++ public/locales/uk.json
@@ -171,7 +171,8 @@
       "confirmDelete": "Ви впевнені, що хочете видалити цю подію?",
       "delete": "Видалити",
       "manageRegistrations": "Керувати реєстраціями",
-      "noGroups": "Групи не знайдені."
+      "noGroups": "Групи не знайдені.",
+      "editEvent": "Редагувати подію"
     },
     "calendarList": {
       "aboutParagraph1": "Календарі дозволяють організовувати та відображати події вашої церкви.",
